fix: move theorist gui to aer.gui
